Platforms: use common POSIXFileBasedLock class
We can refactor and create a common abstract POSIXFileBasedLock
class which can then by shared by both the Linux and Mac
versions, which now only differ in their key internal constants
and type sizes.

From microsoft/VFSForGit@8978fa854f3cb127dda8d872cd490695a08b247a
and microsoft/VFSForGit@7b708506165aaa39783f9bc4d193e9cf39ebc85f.
